Fixed RD-10542
